Async overflow scroll with border-radius renders incorrectly https://bugs.webkit.org/show_bug.cgi?id=194205 <rdar://problem/47771668>
Reviewed by Zalan Bujtas.

Source/WebCore:

When an element has composited overflow:scroll and border-radius, we need to make a layer
to clip to the inside of the border radius if necessary.

Existing code simply turned off needsDescendantsClippingLayer for composited scrolling
layers, but now we check to see if the inner border is rounded. If we have both a m_childContainmentLayer
and scrolling layers, we need to adjust the location of the scrolling layers (which are parented
in m_childContainmentLayer).

Also fix offsetFromRenderer for these layers; it's positive for layers inset from the top left
of the border box.

Tests: compositing/clipping/border-radius-async-overflow-clipping-layer.html
       compositing/clipping/border-radius-async-overflow-non-stacking.html
       compositing/clipping/border-radius-async-overflow-stacking.html

* rendering/RenderLayerBacking.cpp:
(WebCore::RenderLayerBacking::updateConfiguration):
(WebCore::RenderLayerBacking::updateGeometry):
(WebCore::RenderLayerBacking::updateChildClippingStrategy): Layout is always up-to-date now, so remove the comment.

LayoutTests:

New baselines, mostly correcting offsetFromRenderer.

* compositing/clipping/border-radius-async-overflow-clipping-layer-expected.txt: Added.
* compositing/clipping/border-radius-async-overflow-clipping-layer.html: Added.
* compositing/clipping/border-radius-async-overflow-non-stacking-expected.html: Added.
* compositing/clipping/border-radius-async-overflow-non-stacking.html: Added.
* compositing/clipping/border-radius-async-overflow-stacking-expected.html: Added.
* compositing/clipping/border-radius-async-overflow-stacking.html: Added.
* compositing/scrolling/overflow-scrolling-layers-are-self-painting-expected.txt:
* platform/ios/compositing/overflow/scrolling-without-painting-expected.txt:
* platform/ios/compositing/overflow/textarea-scroll-touch-expected.txt: html.css specifies a border-radius on <textarea> for iOS, so we make additional
clipping layers.
* platform/ios/compositing/rtl/rtl-scrolling-with-transformed-descendants-expected.txt:
* platform/ios/compositing/scrolling/overflow-scrolling-layers-are-self-painting-expected.txt:

Identifier: 208701@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@240941 268f45cc-cd09-0410-ab3c-d52691b4dbfc