Web Inspector: "Worker not found" uncaught protocol errors https://bugs.webkit.org/show_bug.cgi?id=194319
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2019-02-06
Reviewed by Matt Baker.

* UserInterface/Base/Main.js:
(WI.terminatePageTarget):
Remove Worker targets associated with the Page on navigation.
Eventually we will have to associate Workers with their parent
frame, but for now a transition can only happen at the Page
level, so we can eliminate all workers on the page.

* UserInterface/Controllers/WorkerManager.js:
(WI.WorkerManager.prototype.workerCreated):
* UserInterface/Protocol/Connection.js:
(InspectorBackend.WorkerConnection.sendMessageToBackend):
Allow any WorkerAgent domain message to fail silently. This can
happen if a Worker is created and destroyed before the frontend
hears about it and sends messages to the backend for that Worker.

Identifier: 208791@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@241039 268f45cc-cd09-0410-ab3c-d52691b4dbfc