iOS Safari will not load USDZ in QLPreview when opening as BLOB https://bugs.webkit.org/show_bug.cgi?id=194766 <rdar://problem/42769186>
Reviewed by Wenson Hsieh.

A download from a Blob URL can happen effectively instantly, which means
that the loadHandler on the registerItemForTypeIdentifier method might not have
been called yet, and we haven't been told what the completionHandler is.
In this case, keep a record of the URL we finished with, and call the completionHandler
right away.

* UIProcess/Cocoa/SystemPreviewControllerCocoa.mm: Add a new private member variable: _downloadedURL.
(-[_WKPreviewControllerDataSource previewController:previewItemAtIndex:]): Call the completionHandler
right away if we've finished.
(-[_WKPreviewControllerDataSource finish:]): Remember that we've finished by recording the URL.

Identifier: 209161@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@241730 268f45cc-cd09-0410-ab3c-d52691b4dbfc