REGRESSION (240698): Fixed position banners flicker and move when scrolling on iOS https://bugs.webkit.org/show_bug.cgi?id=194889 rdar://problem/47755552
Reviewed by Tim Horton.

After r240698 we could commit scrolling changes for a fixed node where the "viewportRectAtLastLayout" and the layer
position were mismatched; this happened when AsyncScrollingCoordinator::reconcileScrollingState() came back from the UI process
with an unstable update and set a new layoutViewport, then some other layout triggered a compositing tree update. During the tree
update, we'd update the fixed scrolling node with the new viewport, and an old layer position.

Fix by ensuring that we only update the geometry info for a scrolling tree node when we update layer geometry for the corresponding
layer.

Not currently testable.

* rendering/RenderLayerCompositor.cpp:
(WebCore::RenderLayerCompositor::updateBackingAndHierarchy):


Identifier: 209260@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@241860 268f45cc-cd09-0410-ab3c-d52691b4dbfc